Create Panel class for the Config

+++ Source/PanelConsole.cpp
@@ -23,7 +23,7 @@ void PanelConsole::Draw()
 	}
 
 	ImGui::SetNextWindowSize(ImVec2(550, 200), ImGuiCond_Always);
-	ImGui::SetNextWindowPos(ImVec2(0, App->window->getCurrentHeight() - 200), ImGuiCond_Always);
+	ImGui::SetNextWindowPos(ImVec2(0, (float)App->window->getCurrentHeight() - 200), ImGuiCond_Always);
 	//ImGui::SetNextWindowBgAlpha(0.75f);
 	ImGui::Begin("Console Log", &visible);
